chg: [user_managemant] clean code + check password and email length
