+++ var/www/modules/settings/templates/users_list.html
@@ -100,7 +100,7 @@
 
 <script>
 $(document).ready(function(){
-  $("#nav_edit_profile").addClass("active");
+  $("#nav_users_list").addClass("active");
 	$("#nav_user_management").removeClass("text-muted");
 } );
 </script>
