+++ var/www/templates/change_password.html
@@ -91,6 +91,10 @@
 						 Digits: 0-9
 						 <span class="badge badge-primary badge-pill">2</span>
 					 </li>
+					 <li class="list-group-item d-flex justify-content-between align-items-center">
+						 Maximum length
+						 <span class="badge badge-primary badge-pill">100</span>
+					 </li>
 					</ul>
 
 		    </form>
